new setup for maven
