--- src/com/example/tests/GoogleSeleniumWebdriverTestCase.java
@@ -1,48 +1,0 @@
-package com.example.tests;
-
-import java.util.regex.Pattern;
-import java.util.concurrent.TimeUnit;
-import org.junit.*;
-import static org.junit.Assert.*;
-import static org.hamcrest.CoreMatchers.*;
-import org.openqa.selenium.*;
-import org.openqa.selenium.firefox.FirefoxDriver;
-import org.openqa.selenium.support.ui.Select;
-
-public class GoogleSeleniumWebdriverTestCase {
-	private WebDriver driver;
-	private String baseUrl;
-	private StringBuffer verificationErrors = new StringBuffer();
-	@Before
-	public void setUp() throws Exception {
-		driver = new FirefoxDriver();
-		baseUrl = "http://www.google.de/";
-		driver.manage().timeouts().implicitlyWait(30, TimeUnit.SECONDS);
-	}
-
-	@Test
-	public void testGoogleSeleniumWebdriverTestCase() throws Exception {
-		driver.get(baseUrl + "/");
-		driver.findElement(By.id("gbqfq")).clear();
-		driver.findElement(By.id("gbqfq")).sendKeys("selenium hq");
-		// ERROR: Caught exception [ERROR: Unsupported command [isTextPresent]]
-	}
-
-	@After
-	public void tearDown() throws Exception {
-		driver.quit();
-		String verificationErrorString = verificationErrors.toString();
-		if (!"".equals(verificationErrorString)) {
-			fail(verificationErrorString);
-		}
-	}
-
-	private boolean isElementPresent(By by) {
-		try {
-			driver.findElement(by);
-			return true;
-		} catch (NoSuchElementException e) {
-			return false;
-		}
-	}
-}
